LayoutTests/editing/pasteboard/data-transfer-items-drop-file-promise-expected.txt


*** Handling drop ***

Types => data
Files => ""
Item list
[0] => (name = 'apple.gif', size = 1476 bytes, type = 'image/gif')
File list
[0] => (name = 'apple.gif', size = 1476 bytes, type = 'image/gif')
